more marketplace stuff

+++ app/src/main/java/com/example/guwap/entity/Item.java
@@ -2,22 +2,22 @@ package com.example.guwap.entity;
 
 //Everything in this class was made by jorge.
 public class Item {
-    private double price;
+    private int price;
     private String name;
     private int quantity;
 
     public Item(Difficulty difficulty, TypeOfItem typeOfItem) {
         if (difficulty == Difficulty.BEGINNER) {
-            this.price = typeOfItem.getPrice() / 2.0;
+            this.price = typeOfItem.getPrice() / 2;
             this.quantity = 0;
         } else if (difficulty == Difficulty.NORMAL) {
             this.price = typeOfItem.getPrice();
             this.quantity = 0;
         } else if (difficulty == Difficulty.HARD) {
-            this.price = typeOfItem.getPrice() * 2.0;
+            this.price = typeOfItem.getPrice() * 2;
             this.quantity = 0;
         } else if (difficulty == Difficulty.IMPOSSIBLE) {
-            this.price = typeOfItem.getPrice() * 4.0;
+            this.price = typeOfItem.getPrice() * 4;
             this.quantity = 0;
         }
         this.name = typeOfItem.getName();
@@ -25,16 +25,16 @@ public class Item {
 
     public Item(Difficulty d, Resources r, PeopleType pt, TypeOfItem typeOfItem) {
         if (d == Difficulty.BEGINNER) {
-            this.price = typeOfItem.getPrice() / 2.0;
+            this.price = typeOfItem.getPrice() / 2;
             this.quantity = (int) (Math.random() * ((400 - 200) + 1)) + 200;
         } else if (d == Difficulty.NORMAL) {
             this.price = typeOfItem.getPrice();
             this.quantity = (int) (Math.random() * ((300 - 100) + 1)) + 100;
         } else if (d == Difficulty.HARD) {
-            this.price = typeOfItem.getPrice() * 2.0;
+            this.price = typeOfItem.getPrice() * 2;
             this.quantity = (int) (Math.random() * ((200 - 0) + 1)) + 0;
         } else if (d == Difficulty.IMPOSSIBLE) {
-            this.price = typeOfItem.getPrice() * 4.0;
+            this.price = typeOfItem.getPrice() * 4;
             this.quantity = (int) (Math.random() * ((100 - 0) + 1)) + 0;
         }
         this.name = typeOfItem.getName();
@@ -50,7 +50,7 @@ public class Item {
         return quantity;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 
@@ -58,7 +58,7 @@ public class Item {
         return name;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
